+++ requests/models.py
@@ -552,13 +552,16 @@ class PreparedRequest(RequestEncodingMixin, RequestHooksMixin):
         """Prepares the given HTTP cookie data.
 
         This function eventually generates a ``Cookie`` header from the
-        given cookies using cookielib. Due to cookielib's design, the header
+        given cookies using cookielib. If `cookies` is None, no ``Cookie``
+        header will be added.  Due to cookielib's design, the header
         will not be regenerated if it already exists, meaning this function
         can only be called once for the life of the
         :class:`PreparedRequest <PreparedRequest>` object. Any subsequent calls
         to ``prepare_cookies`` will have no actual effect, unless the "Cookie"
         header is removed beforehand.
         """
+        if cookies is None:
+            return
         if isinstance(cookies, cookielib.CookieJar):
             self._cookies = cookies
         else: